+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## next
+
+- bug fixes
+    - rare issue that made the last widget(s) in a container disappear when other widgets in the same container had a specific property dependency pattern
+
 ## 1.25.6
 
 - bug fixes